backend/requirements.txt
# DeepFusionColor - Gerekli Python Paketleri
# ============================================
# Kurulum: pip install -r requirements.txt

# Temel Kütüphaneler
numpy>=1.21.0
scipy>=1.7.0

# Görüntü İşleme
opencv-python>=4.5.0
Pillow>=8.3.0
scikit-image>=0.18.0

# Deep Learning (PyTorch)
torch>=1.10.0
torchvision>=0.11.0

# Machine Learning
scikit-learn>=0.24.0

# Wavelet Dönüşümü
PyWavelets>=1.1.1

# Web Framework (API için)
flask>=2.0.0
flask-cors>=3.0.10

# Utilities
matplotlib>=3.4.0
tqdm>=4.62.0
